requirements.txt
bcrypt==4.2.1
blinker==1.9.0
cffi==1.17.1
click==8.1.7
cryptography==44.0.0
Flask==3.1.0
Flask-JWT-Extended==4.7.1
flask-mysql-connector==1.1.0
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
mysqlclient==2.2.6
numpy==2.1.3
pandas==2.2.3
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.1.3
